requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
langchain-experimental>=0.0.50
pandas>=2.0.0
matplotlib>=3.7.0
plotly>=5.15.0
python-dotenv>=1.0.0
tabulate>=0.9.0
statsmodels>=0.14.0
scikit-learn>=1.3.0